feat: add choice to run app with load balancer or without
