+++ package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "./src/index.ts",
   "scripts": {
-    "start:dev": "ts-node-dev ./src/index.ts"
+    "start:dev": "ts-node-dev ./src/index.ts",
+    "start:dev:multi": "ts-node-dev ./src/index.ts -- --multi"
   },
   "keywords": [],
   "author": "",
