+++ src/balancer/balancer.ts
@@ -1,0 +1,73 @@
+import { Worker } from "worker_threads";
+import * as os from "os";
+import { Crud } from "../crud/crud";
+import * as path from "path";
+import * as cluster from "cluster";
+import { URL } from "url";
+import * as http from "http";
+import * as process from "process";
+
+export class Balancer {
+  serverIndex: number;
+
+  constructor(port: number) {
+    const cpus = os.availableParallelism();
+    const workers: { worker: cluster.Worker; port: number }[] = [];
+
+    this.serverIndex = 0;
+
+    if (cluster.default.isPrimary) {
+      for (let i = 0; i < cpus; i++) {
+        const PORT = port + (i + 1);
+        const fork = cluster.default.fork({ PORT });
+        workers.push({ port: PORT, worker: fork });
+      }
+
+      const server = http.createServer((req, res) => {
+        const serverUrl = `http://localhost:${workers[this.serverIndex].port}`;
+        const parsedUrl = new URL(serverUrl);
+
+        if (
+          !parsedUrl.protocol ||
+          !parsedUrl.port ||
+          !parsedUrl.hostname ||
+          !parsedUrl.pathname
+        ) {
+          return;
+        }
+
+        console.log("balancer url", req.url);
+
+        const proxyReq = http.request(
+          {
+            protocol: parsedUrl.protocol,
+            port: parsedUrl.port,
+            host: parsedUrl.hostname,
+            path: req.url,
+            method: req.method,
+          },
+          (proxyRes) => {
+            if (!proxyRes.statusCode) {
+              return;
+            }
+            res.writeHead(proxyRes.statusCode, proxyRes.headers);
+            proxyRes.pipe(res);
+          }
+        );
+
+        req.pipe(proxyReq);
+
+        this.serverIndex = (this.serverIndex + 1) % workers.length;
+      });
+
+      server.listen(8080, () => {
+        console.log("balancer is listening");
+      });
+    } else {
+      if (!process.env.PORT) {
+        return;
+      }
+      const crud = new Crud(parseInt(process.env.PORT));
+    }
+  }
+}
