+++ src/crud/crud.ts
@@ -1,15 +1,25 @@
 import * as http from "http";
 import { Users } from "../users/users";
 
-const users = new Users();
+export class Crud {
+  server: http.Server;
+  users: Users;
 
-export const server = http.createServer((req, res) => {
-  console.log(server.address(), req.url, "request!");
+  constructor(public port: number) {
+    this.users = new Users();
+    this.server = http.createServer((req, res) => {
+      console.table(this.server.address());
+      console.log(req.url, "request!");
 
-  if (req.url?.startsWith("/api/users")) {
-    users.handleRequest(req, res);
-    return;
+      if (req.url?.startsWith("/api/users")) {
+        this.users.handleRequest(req, res);
+        return;
+      }
+      res.writeHead(404);
+      res.end("not found");
+    });
+    this.server.listen(port, () => {
+      console.log("shashlik ya", process.pid, this.port, this.server.address());
+    });
   }
-  res.writeHead(404);
-  res.end("not found");
-});
+}
